% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authority_labs.R
\name{extract_url_data}
\alias{extract_url_data}
\title{Extract URL Data}
\usage{
extract_url_data(var, list_data)
}
\arguments{
\item{var}{a variable}

\item{list_data}{a list}
}
\value{
A character vector
}
\description{
Helper function - extract a variable from list data
}
\examples{
\\donotrun{
extract_url_data(var, list_data)
}
}

