% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authority_labs.R
\name{get_account_data}
\alias{get_account_data}
\title{Get account data}
\usage{
get_account_data(account_id = "331", auth_token = "LmJWkiGmVjQwA1JIIx3o")
}
\arguments{
\item{account_id}{Account number}

\item{auth_token}{API key}
}
\value{
List with response data
}
\description{
Get account data
}
\examples{
get_account_data()
}

