% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authority_labs.R
\name{get_authority_labs_data}
\alias{get_authority_labs_data}
\title{Get data, using callback url}
\usage{
get_authority_labs_data(json_callback_url,
  auth_token = "LmJWkiGmVjQwA1JIIx3o")
}
\arguments{
\item{json_callback_url}{URL from callback}

\item{auth_token}{API key}
}
\value{
A list containing the content of the response
}
\description{
Get data, using callback url
}
\examples{
raw_data <- get_authority_labs_data("http://api.authoritylabs.com/keywords/get.json?keyword=capital\%20one&rank_date=2016-09-13&locale=en-uk&engine=google&pages_from=false&lang_only=false&geo=&autocorrect=&mobile=false&ppl_id=&papi_id=14738015579796")
}

