% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authority_labs.R
\name{get_url_vars}
\alias{get_url_vars}
\title{Get URL variables}
\usage{
get_url_vars(return_data)
}
\arguments{
\item{return_data}{}
}
\value{
Data Frame
}
\description{
Helper function for converting to data frame
}
\examples{
\\donotrun{
get_url_vars(return_data)
}
}

