% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authority_labs.R
\name{set_callback_authority_labs}
\alias{set_callback_authority_labs}
\title{Set Callback}
\usage{
set_callback_authority_labs(keyword,
  callback_url = "http://mhairihmcneill.com/callback",
  auth_token = "LmJWkiGmVjQwA1JIIx3o", locale = "en-uk", mobile = FALSE)
}
\arguments{
\item{keyword}{Keyword we want to analysis}

\item{callback_url}{Callback URL where we can get the results}

\item{auth_token}{API Key}

\item{locale}{Locale for request, as defined in documenation}

\item{mobile}{Set to TRUE to get mobile results}
}
\value{
A list with the response
}
\description{
Set Callback
}
\examples{

set_callback_authority_labs('capital one', 'http://mhairihmcneill.com/callback')
}

