% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authority_labs.R
\name{to_data_frame_authority_labs}
\alias{to_data_frame_authority_labs}
\title{Convert list returned from authority
labs to data frame}
\usage{
to_data_frame_authority_labs(return_data)
}
\arguments{
\item{return_data}{A list from `get_authority_labs_data`}
}
\value{
A dataframe with one row per url
}
\description{
Convert list returned from authority
labs to data frame
}
\examples{
raw_data <- get_authority_labs_data("http://api.authoritylabs.com/keywords/get.json?keyword=mhairi&rank_date=2016-09-13&locale=en-uk&engine=google&pages_from=false&lang_only=false&geo=&autocorrect=&mobile=false&ppl_id=&papi_id=14737953718458")
to_data_frame_authority_labs(raw_data)
}

